buffer-cursor: convert BufferCursor to BufferReader
This changes BufferCursor to be a single purpose reader only class.
