+++ packages/buffer-cursor/lib/BufferReader.ts
@@ -1,5 +1,9 @@
 import assert from "assert";
-export class BufferCursor {
+
+/**
+ * BufferReader class is used to simplify reading information from a Buffer
+ */
+export class BufferReader {
     /**
      * Returns the number of bytes that will be used to encode
      * a BigSize number. BigSize is defined in Lightning Network BOLT 07
@@ -16,7 +20,7 @@ export class BufferCursor {
     private _lastReadBytes: number;
 
     /**
-     * Constructs a read and write cursor for a buffer.
+     * Constructs a reader from the supplied Buffer
      */
     constructor(buffer: Buffer) {
         assert(Buffer.isBuffer(buffer), "Requires a buffer");
@@ -245,118 +249,6 @@ export class BufferCursor {
     }
 
     /**
-     * Write at the current positiion
-     * @param val
-     */
-    public writeUInt8(val: number) {
-        this._writeStandard(this.writeUInt8.name, val, 1);
-    }
-
-    /**
-     * Write at the current positiion
-     * @param val
-     */
-    public writeUInt16LE(val: number) {
-        this._writeStandard(this.writeUInt16LE.name, val, 2);
-    }
-
-    /**
-     * Write at the current positiion
-     * @param val
-     */
-    public writeUInt16BE(val: number) {
-        this._writeStandard(this.writeUInt16BE.name, val, 2);
-    }
-
-    /**
-     * Write at the current positiion
-     * @param val
-     */
-    public writeUInt32LE(val: number) {
-        this._writeStandard(this.writeUInt32LE.name, val, 4);
-    }
-
-    /**
-     * Write at the current positiion
-     * @param val
-     */
-    public writeUInt32BE(val: number) {
-        this._writeStandard(this.writeUInt32BE.name, val, 4);
-    }
-
-    /**
-     * Write at the current positiion
-     * @param value
-     */
-    public writeUInt64LE(value: number | bigint) {
-        const val = BigInt(value);
-        if (val < 0 || val >= BigInt(2) ** BigInt(64)) {
-            throw new RangeError(
-                `The value of "value" is out of range. It must be >= 0 and <= 18446744073709551615. Received ${value.toString()}`,
-            );
-        }
-        const buf = Buffer.from(val.toString(16).padStart(16, "0"), "hex");
-        this.writeBytes(buf.reverse());
-    }
-
-    /**
-     * Write at the current positiion
-     * @param value
-     */
-    public writeUInt64BE(value: number | bigint) {
-        const val = BigInt(value);
-        if (val < 0 || val >= BigInt(2) ** BigInt(64)) {
-            throw new RangeError(
-                `The value of "value" is out of range. It must be >= 0 and <= 18446744073709551615. Received ${value.toString()}`,
-            );
-        }
-        const buf = Buffer.from(val.toString(16).padStart(16, "0"), "hex");
-        this.writeBytes(buf);
-    }
-
-    /**
-     * Write bytes at the current positiion
-     * @param buffer
-     */
-    public writeBytes(buffer: Buffer) {
-        if (!buffer || !buffer.length) return;
-        if (this._position + buffer.length > this._buffer.length) {
-            throw new RangeError("Index out of range");
-        }
-        buffer.copy(this._buffer, this._position);
-        this._position += buffer.length;
-    }
-
-    /**
-     * Reads a variable length unsigned integer as specified in the Lightning Network
-     * protocol documentation and always returns a BigInt to maintain a consistent
-     * call signature.
-     *
-     * @remarks
-     * Specified in:
-     * https://github.com/lightningnetwork/lightning-rfc/blob/master/01-messaging.md#appendix-a-bigsize-test-vectors
-     *
-     * < 0xfd = 1 byte number
-     *   0xfd = 2 byte number (3 bytes total)
-     *   0xfe = 4 byte number (5 bytes total)
-     *   0xff = 8 byte number (9 bytes total)
-     */
-    public writeBigSize(num: bigint) {
-        if (num < BigInt(0xfd)) {
-            this.writeUInt8(Number(num));
-        } else if (num < BigInt(0x10000)) {
-            this.writeUInt8(0xfd);
-            this.writeUInt16BE(Number(num));
-        } else if (num < BigInt(0x100000000)) {
-            this.writeUInt8(0xfe);
-            this.writeUInt32BE(Number(num));
-        } else {
-            this.writeUInt8(0xff);
-            this.writeUInt64BE(num);
-        }
-    }
-
-    /**
      * Helper for reading off buffer using built-in read functions
      * @param fn name of function
      * @param len length to read
@@ -369,20 +261,5 @@ export class BufferCursor {
         this._position += len;
         this._lastReadBytes = len;
         return result;
-    }
-
-    /**
-     * Helper for writing to the buffer using built-in write
-     * functions
-     * @param fn name of function
-     * @param val number to write
-     * @param len length of number in bytes
-     */
-    private _writeStandard(fn: string, val: number, len: number) {
-        if (this._position + len > this._buffer.length) {
-            throw new RangeError("Index out of range");
-        }
-        this._buffer[fn](val, this._position);
-        this._position += len;
     }
 }
