+++ packages/buffer-cursor/lib/index.ts
@@ -2,6 +2,6 @@ export * from "./bigToBufLE";
 export * from "./bigToBufBE";
 export * from "./bufToStream";
 export * from "./encodeVarInt";
-export * from "./BufferCursor";
+export * from "./BufferReader";
 export * from "./BufferWriter";
 export * from "./StreamReader";
